Almost complete, commiting so that linking after submitting can be done by those who have access to database.
Need to update js to react since just importing react cause all functionality to stop working, also working on finding a way to get character count to work inside the text box for coments.

Managed to fix error where date for expiration wouldn't appear when selecting no if expiration date doesn't match what was advertised.

Also updated nav links to match for when we separate accounts between user and organization as well as linking to home. Besides this, I also added a small change to nav to prevent it from being scrollable and always keeping it on the page