Donate Form Created the donate form
Front End is complete

Java Script code for unique ID generation needs to be worked on. To ensure id truly is unique, the id needs to be crosscheck with id's stored in the data base, and if it doesn't exist stored in the database, else needs to form a new id.

Honestly really not sure if the js code that i created will work or not.

For the logo, should appear once merged into frontend since png file is in other branchs and will be merged also.